refactor utils

+++ services/LevelManager.h
@@ -10,6 +10,7 @@
 #define LevelManager_h
 
 #include <vector>
+#include "../utils/Array.h"
 
 #include "../Debug.h"
 #include "../interfaces/TypeDefs.h"
@@ -44,7 +45,7 @@ class LevelManager: public BundleManager {
     LevelManager(SpriteManager& spriteManager, ResourceManager& resourceManager, Screen& screen);
     bool initialize() override;
     
-    unsigned int getRealmsCount();
+    unsigned long getRealmsCount();
     void getRealm(LevelRealm& realm, unsigned int realmId);
     LevelRealm getRealm(unsigned int realmId);
